Use a more robust check for process completion
The approach used in the script assumes that the GPU is exclusively used
by the invoker. However, some places (like mine) treat GPUs as shared
resources. Therefore, the single presence of a process called `python`
in the process list, even if not created by the caller, will cause the
script to loop needlessly.

This change creates an array where the PIDs for each of the calculations
are stored, then uses `wait` to ensure that all the processes have ended
before moving to the next iteration.

**Note**: This adds a dependency on bash (but I'm not sure if this
script is POSIX-compatible) and requires a reasonably recent version of
it (tested with 5.0 on Debian).